# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CSI Portal API
  version: 1.1.0
  description: |
    Current API contract for CSI Web App.
    Notes:
    - User endpoints accept `UserId` or `Username` where the path uses `{id}`.
    - Survey/Event endpoints accept `SurveyId` (UUID) or `SurveyNo` (numeric string).
    - Event routes are aliases of survey routes and are used by the current frontend.
servers:
  - url: http://localhost:3000
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: adam.cid00676 }
        password: { type: string, example: admin123 }
    AuthUser:
      type: object
      properties:
        userId: { type: string, format: uuid }
        userKey: { type: string, example: adam.cid00676 }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [SuperAdmin, AdminEvent, ITLead, DepartmentHead]
    LoginResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        user: { $ref: "#/components/schemas/AuthUser" }
...